+++ game_manager/board_manager.py
@@ -263,7 +263,7 @@ class BoardData(object):
         if self.tryMoveCurrent(self.currentDirection, self.currentX - 1, self.currentY):
             self.currentX -= 1
         else:
-            print("failed to moveLeft..")
+            #print("failed to moveLeft..")
             return False
         return True
 
@@ -271,7 +271,7 @@ class BoardData(object):
         if self.tryMoveCurrent(self.currentDirection, self.currentX + 1, self.currentY):
             self.currentX += 1
         else:
-            print("failed to moveRight..")
+            #print("failed to moveRight..")
             return False
         return True
 
@@ -280,7 +280,7 @@ class BoardData(object):
             self.currentDirection += 1
             self.currentDirection %= 4
         else:
-            print("failed to rotateRight..")
+            #print("failed to rotateRight..")
             return False
         return True
 
@@ -289,7 +289,7 @@ class BoardData(object):
             self.currentDirection -= 1
             self.currentDirection %= 4
         else:
-            print("failed to rotateLeft..")
+            #print("failed to rotateLeft..")
             return False
         return True
