+++ start_train.sh
@@ -1,0 +1,2 @@
+
+python start.py -m train -d 50 -t -1
